Get rid of spawn injection
Respawn system likes to throw player out of his spawn, while death occured during movement.

+++ 19in1.h
@@ -1,10 +1,10 @@
 #pragma once
-#include <vadefs.h>
+#include <stdint.h>
 
 extern bool unlockMaps;
 extern void listMaps();
 
 constexpr uintptr_t MpMapTableBaseAddr = 0x757B28;
-constexpr size_t MpMapRecordSize = 0x28;
-constexpr size_t MpMapRecordCount = 34;
-constexpr size_t MpMapRecordFlagsOffset = 0x20;
+constexpr uint8_t MpMapRecordSize = 0x28;
+constexpr uint8_t MpMapRecordCount = 34;
+constexpr uint8_t MpMapRecordFlagsOffset = 0x20;
